feat: 스탯 Modifer 시스템 추가

+++ Assets/_DataTables/Scripts/EntityStatsTableEntry.cs
@@ -22,7 +22,7 @@ namespace DungeonShooter
         public int Defense { get; set; }
 
         /// <summary>이동 속도</summary>
-        public float MoveSpeed { get; set; }
+        public int MoveSpeed { get; set; }
     }
 }
